refactor smooth scroll functionality to use utility function across components
